+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# unreleased
+# 1.11.0
 * Normalize card_faces so all card objects have a card_faces array
 
 _Breaking Changes_